Move meshcat server and recorder process methods in dedicated file to fix 'spawn' support (Win10).
